bitint: Zero-extend the result of a signed->unsigned widening cast
A widening cast from a signed _BitInt operand to an unsigned _BitInt
type involves filling the extra limb(s) with sign extension.
On a target that wants _BitInts extended in memory, if this unsigned
type has a partial limb, the unused part of it should be zeroed.

e.g. Assuming limb_mode == E_DImode, at the end of

  void
  test (unsigned _BitInt(519) *t, _BitInt(512) x)
  {
    *t = -x;
  }

the most significant limb of *t should be masked with 0x7f.

This patch also fixes gcc.dg/torture/bitint-16.c, which aborts at -O2
when the extension on load is optimized away.

	* gimple-lower-bitint.cc (bitint_large_huge::lower_mergeable_stmt):
	Zero-extend the partial limb of any unsigned _BitInt LHS assigned
	with a widening sign-extension.

	* gcc.dg/torture/bitint-83.c: New test.