Darwin: Anchor block internal symbols must not be linker-visible.
When we are using section anchors, there's a requirement that the
sequence of the content is an unbroken block.  If we allow linker-
visible symbols in that block, ld(64) would be able to break it
into sub-sections on those symbol boundaries.

Do not allow symbols that should be visible to be anchored.
Do not make anchor block internal symbols linker-visible.

gcc/ChangeLog:

	* config/darwin.cc (darwin_encode_section_info): Do not
	make anchored symbols linker-visible.
	(darwin_use_anchors_for_symbol_p): Disallow anchoring on
	symbols that must be linker-visible (or external), even
	if the definitions are in this TU.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>