Darwin: Handle string constants specially when asan is enabled.
The Darwin ABI uses a different section for string constants when
address sanitizing is enabled.  This adds defintions of the asan-
specific sections and switches string constants to the correct
section.

It also makes the string constant symbols linker-visible when
asan is enabled, but not otherwise.

gcc/ChangeLog:

	* config/darwin-sections.def (asan_string_section,
	asan_globals_section, asan_liveness_section): New.
	* config/darwin.cc (objc_method_decl): Use asan sections
	when asan is enabled.
	(darwin_encode_section_info): Alter string constant
	linker visibility depending on asan.
	(machopic_select_section): Use the asan sections when
	asan is enabled.

gcc/testsuite/ChangeLog:

	* gcc.dg/torture/darwin-cfstring-3.c: Adjust for amended
	string labels.
	* g++.dg/torture/darwin-cfstring-3.C: Likewise.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>