Add -mgrow-frame-downwards
Grow the local frame down instead of up for mips16 code size.

By growing the frame downwards we get spill slots created at the lowest
address rather than highest address in a local frame. The benefit being
that when the frame is large the spill slots can still be accessed using
a 16bit instruction whereas it is less important for large local
variables to be accessed using short instructions as they are (probably)
accessed less frequently.

This is default on for MIPS16.

gcc/
	* config/mips/mips.h (FRAME_GROWS_DOWNWARD) Allow the frame to
	grow downwards for mips16 when -mgrow-frame-downwards is set.
	* config/mips/mips.opt: Add -mgrow-frame-downwards option.