tree-optimization/121493 - another missed VN with aggregate copy
This is another case where opportunistically handling a first
aggregate copy where we failed to match up the refs exactly
(as we don't insert missing handling components) yields to a
failure in the second aggregate copy that we visit.  Add another
fixup to deal with such situations, in-line with that present
opportunistic handling.

	PR tree-optimization/121493
	* tree-ssa-sccvn.cc (vn_reference_lookup_3): Opportunistically
	strip components with known offset.

	* gcc.dg/tree-ssa/ssa-fre-109.c: New testcase.