ada: Ignore Ghost policy checks for some pragmas
Some pragmas where the ghostness is based on the argument of
the pragma triggered ghost policy errors because we had not marked
the pragmas as ghost. However we could only do so once the argument
of the pragma was analyzed.

We can safely ignore the policy checks for those pragmas since
if the argument was ghost then the pragma also had to be ghost.

Marking the pragma afterwards as ghost is only for the cleanup of
ignored ghost purposes.

gcc/ada/ChangeLog:

	* ghost.adb (Is_OK_Pragma): Use the
	Suppressed_Ghost_Policy_Check_Pragma list for ignoring certain
	pragmas.
	* sem_prag.ads (Suppressed_Ghost_Policy_Check_Pragma): New variable
	to store the pragmas that could be ignored when checking for
	consitant ghost policy.