D: Adjust the code-gen for a string constant.
In this function, we are generating a string constant but do so with
a mismatch between the actual string length and the length specified
in the type.  This causes Darwin, at least, to place the string in an
unexpected section (since the parameters do not match, it is rejected
as a cstring).

Use build_string_literal() to construct a consistent null-terminated
string.

gcc/d/ChangeLog:

	* d-codegen.cc (build_filename_from_loc): Use
	build_string_literal() to build a null-terminated string for
	the filename.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>